Move Components to separate files
