+++ lib/components/result-view/index.js
@@ -6,7 +6,7 @@ import { action, observable } from "mobx";
 import { createImmutableOutput } from "@nteract/commutable";
 
 import { reactFactory } from "./../../utils";
-import ResultViewComponent from "./component";
+import ResultViewComponent from "./result-view";
 
 import type { IObservableArray, IObservableValue } from "mobx";
